Pull request 565: Add an option to build a 64-bit library with MSVC on Windows
Merge in ADGUARD-CORE-LIBS/dns-libs from feature/windows_msvc_x64_library to master

Squashed commit of the following:

commit 54e5519f7843f110fd0d085693061119c07adf4f
Merge: 302c525b c81b60e3
Author: Sergey Fionov <sfionov@adguard.com>
Date:   Fri Nov 24 08:37:19 2023 +0200

    Merge remote-tracking branch 'origin/master' into feature/windows_msvc_x64_library

commit 302c525b6202e8c303c821deb5e3a61bbd730076
Author: Danila-Arg <d.argastsev@adguard.com>
Date:   Tue Nov 21 19:33:48 2023 +0300

    fixed nuspec file

commit 6e91e597226c7378f87c515a2d650fdfcf52ed2b
Author: Danila-Arg <d.argastsev@adguard.com>
Date:   Mon Nov 20 16:23:17 2023 +0300

    fixed build

commit 9570eb13699fbfd233742e8df96e8c6eb4eb0eb4
Author: Sergey Fionov <sfionov@adguard.com>
Date:   Mon Nov 20 13:45:42 2023 +0200

    Temporary disable tests

commit 77a016e220ca65e98b2ae10813dc8fc16245a9d7
Author: Danila-Arg <d.argastsev@adguard.com>
Date:   Mon Nov 20 11:53:32 2023 +0300

    Fixed tests

commit deb212eb014a9dd84022b5d7b1177c5f2472c05d
Author: Danila-Arg <d.argastsev@adguard.com>
Date:   Mon Nov 20 10:19:57 2023 +0300

    pr review

commit 82543a30b42212e1d64a823773e301cc05f1b0a6
Author: Danila-Arg <d.argastsev@adguard.com>
Date:   Fri Nov 17 18:56:17 2023 +0300

    Update CMakeLists.txt

commit a382c786e61fff634da962a1523bd02ea10a5f10
Merge: 96e95e46 48e1cc63
Author: Danila-Arg <d.argastsev@adguard.com>
Date:   Fri Nov 17 18:30:06 2023 +0300

    Merge branch 'master' into feature/windows_msvc_x64_library

commit 96e95e46f8198ba745c596aac7412892d8e8ec5e
Author: Danila-Arg <d.argastsev@adguard.com>
Date:   Fri Nov 17 18:29:44 2023 +0300

    self review

commit 6a2592b2f432039d982f69f151aa544cfa759307
Author: Danila-Arg <d.argastsev@adguard.com>
Date:   Fri Nov 17 12:54:02 2023 +0300

    make pretty

commit 2013fd17315b87dc4209ac551b7a16b03fc420be
Author: Danila-Arg <d.argastsev@adguard.com>
Date:   Fri Nov 17 12:51:07 2023 +0300

    self review

commit 99096f458ee439d95d652153070b2a274f42fb48
Author: Danila-Arg <d.argastsev@adguard.com>
Date:   Fri Nov 17 12:50:04 2023 +0300

    make simillar to CoreLibs

commit 4f0d52d846383940a320a878ddc0af61e336ccf1
Author: Boris <boris@adguard.com>
Date:   Tue Nov 14 20:05:48 2023 +0300

    windows_msvc_x64_library, fix calback

commit 11d7012955f1e503d57edcd42987cd4aa6bf9032
Author: Sergey Fionov <sfionov@adguard.com>
Date:   Tue Nov 14 16:05:10 2023 +0200

    Fix

commit 3b2101d43526f7b32fdadef898e8c2e3eaa12a0a
Author: Sergey Fionov <sfionov@adguard.com>
Date:   Tue Nov 14 16:04:05 2023 +0200

    Fix build on Win64

commit ef65befe8f44904107d042e5bd175b75da97aaab
Author: Danila-Arg <d.argastsev@adguard.com>
Date:   Tue Nov 14 16:05:15 2023 +0300

    Update DnsApiConverter.cs

commit 33b36147d8378ae4b0c53ac98138f05ea29c83b1
Author: Danila-Arg <d.argastsev@adguard.com>
Date:   Tue Nov 14 10:07:51 2023 +0300

    Update AGDnsApi.cs

commit 787271db76c03bf68adfa13c461e5fcd8f5247c6
Author: Danila-Arg <d.argastsev@adguard.com>
Date:   Tue Nov 14 10:06:27 2023 +0300

    fixed merge problem

commit 3c7816b3814b50605201a7f6907ee0752f759fa3
Author: Danila-Arg <d.argastsev@adguard.com>
Date:   Mon Nov 13 18:28:47 2023 +0300

    Update DnsLibsDllProvider.cs

commit ef62d0cc9de43762271a150611d06ce439702270
Merge: c09ca53c 36d1c836
Author: Boris <boris@adguard.com>
Date:   Mon Nov 13 16:22:19 2023 +0300

    pull from master

... and 21 more commits

+++ platform/windows/capi/src/ag_dns.cpp
@@ -662,7 +662,7 @@ void ag_dnsproxy_handle_message_async(ag_dnsproxy *handle, ag_buffer c_message, 
     coro::run_detached([](DnsProxy *proxy, Uint8Vector message, std::optional<DnsMessageInfo> info,
                                ag_handle_message_async_cb handler) -> coro::Task<void> {
         auto result = co_await proxy->handle_message({message.data(), message.size()}, opt_as_ptr(info));
-        ag_buffer result_buffer{.data = result.data(), .size = result.size()};
+        ag_buffer result_buffer{.data = result.data(), .size = (uint32_t) result.size()};
         handler(&result_buffer);
     }(proxy, std::move(message), std::move(info), handler));
 }
